refactor(ocpp): use auth barrel index instead of deep imports
